Proper Wheel Rendering, Model Cleanup
Got the wheels working! Only 4 right now since that's what the default snippet came with. But with a bit of messing around with the .json we should be able to get 6. Until then the back 2 are missing.

The wheels roll, and the front ones turn left and right. Its all jank but importantly it works!

Models folder has been overhauled to get rid of test assets and replace them with more proper ones we can update over time.