Made Transform more versatile
Transform now contains linear velocity and forward, up, and right vectors.

Modifying values has been changed to a system of getters and setters. Changing the position or rotation of an object will automatically update its direction vectors.

Linear velocity is (0,0,0) on non-physics objects, but is continually updated for physics objects. For an example, the speed of an object can be calculated by taking the length of the linear velocity vector.